% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{append_exchangeratehost_access_key}
\alias{append_exchangeratehost_access_key}
\title{Retrieves exchangerate.host (forex) API key from R environment variables
and appends to API call}
\usage{
append_exchangeratehost_access_key(url)
}
\arguments{
\item{url}{A URL representing an API endpoint}
}
\value{
The input URL with API access key appended as a URL parameter
}
\description{
Retrieves exchangerate.host (forex) API key from R environment variables
and appends to API call
}
\examples{
\dontrun{
base_Url <- "https://api.exchangerate.host/latest?base=USD"
base_Url \%>\% append_exchangeratehost_access_key

# [1] "https://api.exchangerate.host/latest?base=USD&access_key=7e5e3140140bd8e4f4650cc41fc772c0"

}


}
