% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{convert_to_iso2Code}
\alias{convert_to_iso2Code}
\title{Convert any country input into its iso2Code}
\usage{
convert_to_iso2Code(country_input_type_string, country)
}
\arguments{
\item{country_input_type_string}{Either "country_name" or "iso2Code" - use country_input_type(country, countries_dataframe) to determine or assign manually.}

\item{country}{A country/region name or iso2Code.}
}
\description{
`convert_to_iso2Code` accepts the type of country input and the country, and returns the relevant iso2Code
}
\examples{

# Provide a salary string and 'extract_salary' and will extract the salary and return it
country <- "Australia"
country_input_type_string <- "country_name"
convert_to_iso2Code(country_input_type_string, country)
# [1] "AU"

country <- "AU"
country_input_type_string <- "iso2Code"
convert_to_iso2Code(country_input_type_string, country)
# [1] "AU"


}
