% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{currencies}
\alias{currencies}
\title{Retrieve available currencies and their respective symbols/codes}
\usage{
currencies()
}
\value{
A data.frame of available currencies and their respective symbols/codes
}
\description{
Retrieve available currencies and their respective symbols/codes
}
\examples{
\dontrun{
# Display available currencies and their respective symbols/codes
currencies()
#                                description code
# 1              United Arab Emirates Dirham  AED
# 2                           Afghan Afghani  AFN
# 3                             Albanian Lek  ALL
# 4                            Armenian Dram  AMD
# 5            Netherlands Antillean Guilder  ANG
# 6                           Angolan Kwanza  AOA
# 7                           Argentine Peso  ARS
}


}
