% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{currency_characters}
\alias{currency_characters}
\title{Provide currency characters}
\usage{
currency_characters()
}
\value{
A character vector of currency symbols
}
\description{
Provide currency characters
}
\examples{

currency_characters()

}
