% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{currency_characters_ascii}
\alias{currency_characters_ascii}
\title{Provide currency characters available in ASCII}
\usage{
currency_characters_ascii()
}
\description{
Provide currency characters available in ASCII
}
\examples{

currency_characters_ascii()

}
