% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{currency_info}
\alias{currency_info}
\title{Provides information for 191 currencies}
\usage{
currency_info()
}
\value{
A data.frame containing currency information for 191 currencies.
    Currency information includes: name, iso code, currency symbol (and
    alternative symbols if applicable), subunit, number of subunits per major
    unit, whether the currency symbol ought to appear before or after the
    number of units, display format, html entity, decimal mark,
    thousands separator, iso numeric, and smallest denomination.
}
\description{
Provides information for 191 currencies
}
\examples{
 \dontrun{
# currency_info <- currency_info()
# # Inspect first 6 rows
# head(currency_info)
}
}
