% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{currency_to_numeric}
\alias{currency_to_numeric}
\title{Convert human readable currencies into numeric data}
\usage{
currency_to_numeric(currency_text)
}
\arguments{
\item{currency_text}{Price or vector of prices}
}
\value{
A numeric vector
}
\description{
Convert human readable currencies into numeric data
}
\examples{
library(dplyr)
c("$134,345.05", "£22", "¥30000") \%>\% currency_to_numeric()
# [1] 134345     22  30000
}
