% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{pminmax}
\alias{pminmax}
\title{Removes redundant API calls of currency pairs. That is, revmoes the need
to for separate calls for both
`from = EUR, to = USD` and `from = USD, to = EUR`}
\usage{
pminmax(x, y)
}
\arguments{
\item{x}{A currency code (see currencies() for supported codes)}

\item{y}{A currency code}
}
\value{
A character vector
}
\description{
Removes redundant API calls of currency pairs. That is, revmoes the need
to for separate calls for both
`from = EUR, to = USD` and `from = USD, to = EUR`
}
\examples{

# See: https://stackoverflow.com/a/73254014/9059865



}
