% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceR.R
\docType{package}
\name{priceR-package}
\alias{priceR-package}
\title{priceR: Economics and Pricing Tools}
\description{
priceR: Economics and Pricing Tools
}
\details{
It has the goal of providing a quick and practical way of extracting numeric price and currency data from text fields, thus allowing faster and easier analysis of ostensibly numeric data.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
