% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_nearest.R
\name{round_down_to_nearest}
\alias{round_down_to_nearest}
\title{Round prices down to the nearest specified increment}
\usage{
round_down_to_nearest(amount, to_nearest)
}
\arguments{
\item{amount}{Price to be rounded}

\item{to_nearest}{Increment to which price is to be rounded down to}
}
\description{
Round prices down to the nearest specified increment
}
\examples{

# Round down to nearest 0.05 (5c)
library(dplyr)
prices <- c(4.45, 5.22, 0.16, 27.88, 112.19)
prices \%>\% round_down_to_nearest(0.05)


# Round down to nearest $10
prices <- c(4.45, 5.22, 0.16, 27.88, 112.19)
prices \%>\% round_down_to_nearest(10)
}
