% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{show_countries}
\alias{show_countries}
\title{Show available country codes}
\usage{
show_countries()
}
\description{
`show_countries` calls the World Bank API and retreives a list of available countries and regions
}
\examples{

# Simply call show_countries() to receive a dataframe of all countries (and regions) and their
# iso2Code
show_countries()
    iso2Code         country_name
1         AW                Aruba
2         AF          Afghanistan
3         A9               Africa
4         AO               Angola
Etc

# Assign so as to save on API calls (recommended)
countries_dataframe <- show_countries()

}
