% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\docType{data}
\name{currency_info}
\alias{currency_info}
\title{Information for each of 191 currencies}
\format{
An object of class \code{data.frame} with 191 rows and 15 columns.
}
\usage{
currency_info
}
\value{
A data.frame containing currency information for 191 currencies.
    Currency information includes: name, iso code, currency symbol (and
    alternative symbols if applicable), subunit, number of subunits per major
    unit, whether the currency symbol ought to appear before or after the
    number of units, display format, html entity, decimal mark,
    thousands separator, iso numeric, and smallest denomination.
}
\description{
Information for each of 191 currencies
}
\examples{
head(currency_info)
}
\keyword{datasets}
