% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceR.R
\docType{package}
\name{priceR}
\alias{priceR}
\alias{priceR-package}
\title{priceR: Regular Expressions for Prices and Currencies}
\description{
Applying regular expressions to text data to extract prices and currencies in numeric class, allowing for faster and easier analysis
}
\details{
It has the goal of providing a quick and practical way of extracting numeric price and currency data from text fields, thus allowing faster and easier analysis of ostensibly numeric data.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
